Change call features to call/save only. Add send cmd.
