+++ iwallet/call.go
@@ -17,12 +17,8 @@ package iwallet
 import (
 	"github.com/spf13/cobra"
 
-	"github.com/iost-official/go-iost/ilog"
 	"github.com/iost-official/go-iost/rpc/pb"
-	"github.com/iost-official/go-iost/sdk"
 )
-
-var inputTxFile string
 
 // callCmd represents the call command that call a contract with given actions.
 var callCmd = &cobra.Command{
@@ -33,36 +29,27 @@ var callCmd = &cobra.Command{
 	An ACTION is a group of 3 arguments: contract name, function name, method parameters.
 	The method parameters should be a string with format '["arg0","arg1",...]'.`,
 	Example: `  iwallet call "token.iost" "transfer" '["iost","user0001","user0002","123.45",""]' --account test0
-  iwallet call --tx_file tx.json --account test0`,
+  iwallet call "token.iost" "transfer" '["iost","user0001","user0002","123.45",""]' --output tx.json`,
 	Args: func(cmd *cobra.Command, args []string) error {
-		return checkAccount(cmd)
+		if outputTxFile == "" {
+			return checkAccount(cmd)
+		}
+		return nil
 	},
 	RunE: func(cmd *cobra.Command, args []string) error {
-		trx := &rpcpb.TransactionRequest{}
-		if inputTxFile != "" {
-			if len(args) != 0 {
-				ilog.Warnf("load tx from file %v, will ignore cmd args %v", inputTxFile, args)
-			}
-			err := sdk.LoadProtoStructFromJSONFile(inputTxFile, trx)
-			if err != nil {
-				return err
-			}
-		} else {
-			var actions []*rpcpb.Action
-			actions, err := actionsFromFlags(args)
-			if err != nil {
-				return err
-			}
-			trx, err = initTxFromActions(actions)
-			if err != nil {
-				return err
-			}
+		var actions []*rpcpb.Action
+		actions, err := actionsFromFlags(args)
+		if err != nil {
+			return err
 		}
-		return sendTx(trx)
+		tx, err := initTxFromActions(actions)
+		if err != nil {
+			return err
+		}
+		return saveOrSendTx(tx)
 	},
 }
 
 func init() {
 	rootCmd.AddCommand(callCmd)
-	callCmd.Flags().StringVarP(&inputTxFile, "tx_file", "", "", "load tx from this file")
 }
